requirements.txt
python-docx
Pillow>=10.0.0
